fix: conform to the new scopus format
Scopus changed the format by not providing the Abbreviated Source Title
and changing the format of the authors. We'll use the Source title field
now and the authors seem to come in `initials, last name, first name`
formats.

So we're going to re-format the authors as `last name, initials` as it
was before.